refactor(util): remove function, rename file
